🚚 rename: storybook test를 위해 폴더 구조 변경
