+++ apps/client/src/entities/home/GuideVideo/GuideVideo.stories.tsx
