+++ apps/client/src/entities/home/GuideVideo/GuideVideo.tsx
