+++ apps/client/src/entities/home/HoveredEmptyWorkspace/HoveredEmptyWorkspace.tsx
