+++ apps/client/src/entities/home/NotHoveredEmptyWorkspace/NotHoveredEmptyWorkspace.tsx
