+++ apps/client/src/entities/home/WorkspaceAddBtn/WorkspaceAddBtn.tsx
