+++ apps/client/src/entities/home/WorkspaceItem/WorkspaceItem.tsx
