+++ apps/client/src/entities/home/WorkspaceLoadError/WorkspaceLoadError.stories.tsx
@@ -1,4 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react';
+
 import { WorkspaceLoadError } from './WorkspaceLoadError';
 
 const meta: Meta<typeof WorkspaceLoadError> = {
