+++ apps/client/src/entities/home/WorkspaceLoadError/WorkspaceLoadError.tsx
