+++ apps/client/src/entities/index.ts
@@ -1,13 +1,13 @@
-export { GuideVideo } from './home/GuideVideo';
-export { WorkspaceItem } from './home/WorkspaceItem';
-export { WorkspaceAddBtn } from './home/WorkspaceAddBtn';
-export { WorkspaceLoadError } from './home/WorkspaceLoadError';
+export { GuideVideo } from './home/GuideVideo/GuideVideo';
+export { WorkspaceItem } from './home/WorkspaceItem/WorkspaceItem';
+export { WorkspaceAddBtn } from './home/WorkspaceAddBtn/WorkspaceAddBtn';
+export { WorkspaceLoadError } from './home/WorkspaceLoadError/WorkspaceLoadError';
 
 export { RedoButton } from './workspace/RedoButton/RedoButton';
 export { UndoButton } from './workspace/UndoButton/UndoButton';
 export { SaveButton } from './workspace/SaveButton/SaveButton';
 export { WorkspaceNameInput } from './workspace/WorkspaceNameInput/WorkspaceNameInput';
-export { CssTooltip } from './workspace/CssTooltip';
+export { CssTooltip } from './workspace/CssTooltip/CssTooltip';
 export { CssOptionItem } from './workspace/CssOptionItem/CssOptionItem';
 export { CssCategoryButton } from './workspace/CssCategoryButton/CssCategoryButton';
-export { RenderResetCssTooltip } from './workspace/RenderResetCssTooltip';
+export { RenderResetCssTooltip } from './workspace/RenderResetCssTooltip/RenderResetCssTooltip';
